# internal/docs/swagger.yaml
basePath: /
definitions:
...
  api.authResponse:
    properties:
      tokens:
        $ref: '#/definitions/service.TokenPair'
      twoFactor:
        $ref: '#/definitions/service.Challenge'
      user:
        $ref: '#/definitions/domain.User'
    type: object
  api.changePasswordRequest:
    properties:
      currentPassword:
        type: string
      newPassword:
        type: string
      passkeyResponse:
        items:
          type: integer
        type: array
      passkeySessionId:
        description: |-
          PasskeySessionID and PasskeyResponse are the alternative proof. The
          response is left raw: it is the browser's credential object, which does
          not survive a strict decode.
        type: string
    type: object
...
  api.emailChangeRequest:
    properties:
      newEmail:
        type: string
    type: object
  api.emailRequest:
    properties:
      email:
        type: string
    type: object
...
  api.errorBody:
    properties:
      error:
        type: string
    type: object
...
  api.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  api.refreshRequest:
    properties:
      refreshToken:
        type: string
    type: object
  api.registerRequest:
    properties:
      bootstrapSecret:
        type: string
      email:
        type: string
      locale:
        description: |-
          Locale is the language picked on the form. Optional: an absent or
          unsupported value leaves the account on the default.
        type: string
      timeZone:
        description: |-
          TimeZone is the browser's IANA zone, used before the account can sign in
          and save preferences itself.
        type: string
    type: object
...
  api.resetPasswordRequest:
    properties:
      newPassword:
        type: string
      token:
        type: string
    type: object
...
  api.tokenRequest:
    properties:
      token:
        type: string
    type: object
  api.twoFactorVerifyRequest:
    properties:
      challengeId:
        type: string
      code:
        type: string
    type: object
...
  domain.User:
    properties:
      createdAt:
        type: string
      email:
        description: |-
          Email is the account identity: unique, required, and stored lower-cased.
          There is no separate username — one fewer thing to choose, to remember,
          and to have to police for impersonation.
        type: string
      emailVerifiedAt:
        description: |-
          EmailVerifiedAt is nil until the address is proven. Nullable, so it can
          be added to a populated table.
        type: string
      id:
        type: integer
      isAdmin:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  service.Challenge:
    properties:
      challengeId:
        type: string
      expiresAt:
        type: string
    type: object
...
  service.TokenPair:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: string
      refreshToken:
        type: string
    type: object
  service.Usage:
    properties:
      contextLimit:
        type: integer
      contexts:
        type: integer
      noteLimit:
        type: integer
      notes:
        type: integer
      projectLimit:
        type: integer
      projects:
        type: integer
      recurring:
        type: integer
      recurringLimit:
        type: integer
      storageBytes:
        type: integer
      storageLimit:
        type: integer
      tagLimit:
        type: integer
      tags:
        type: integer
      todoLimit:
        type: integer
      todos:
        type: integer
    type: object
info:
  contact: {}
  description: 'A GTD web application: actions, contexts, projects and the tickler.'
  title: gotracks API
  version: "1.0"
paths:
...
  /api/v1/auth/2fa/verify:
    post:
      parameters:
      - description: Challenge id and code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.twoFactorVerifyRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Complete a two-factor sign-in
      tags:
      - auth
  /api/v1/auth/account/deletion/confirm:
    post:
      parameters:
      - description: Mailed deletion token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.tokenRequest'
      responses:
        "204":
          description: account deleted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
        "409":
          description: last administrator
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Confirm permanent account deletion
      tags:
      - auth
  /api/v1/auth/email/change/confirm:
    post:
      parameters:
      - description: Mailed email-change token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.tokenRequest'
      responses:
        "204":
          description: email changed
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Confirm a new email address
      tags:
      - auth
  /api/v1/auth/email/resend:
    post:
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.emailRequest'
      responses:
        "204":
          description: accepted
      summary: Resend the verification email
      tags:
      - auth
  /api/v1/auth/email/verify:
    post:
      parameters:
      - description: Mailed token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.tokenRequest'
      responses:
        "204":
          description: verified
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Verify an email address
      tags:
      - auth
  /api/v1/auth/invitation/accept:
    post:
      parameters:
      - description: Invitation token and password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.resetPasswordRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Accept a user invitation
      tags:
      - auth
  /api/v1/auth/login:
    post:
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Sign in with email and password
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.refreshRequest'
      responses:
        "204":
          description: signed out
      summary: Sign out (revoke a refresh token)
      tags:
      - auth
...
  /api/v1/auth/password/forgot:
    post:
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.emailRequest'
      responses:
        "204":
          description: accepted
      summary: Request a password-reset email
      tags:
      - auth
  /api/v1/auth/password/reset:
    post:
      parameters:
      - description: Token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.resetPasswordRequest'
      responses:
        "204":
          description: password changed
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Reset a password from a mailed link
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.refreshRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TokenPair'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Rotate a refresh token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      parameters:
      - description: Email and optional locale
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.registerRequest'
      responses:
        "204":
          description: accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
        "403":
          description: registration disabled
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Request account enrollment
      tags:
      - auth
...
  /api/v1/me:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Get the current account
      tags:
      - account
  /api/v1/me/deletion:
    post:
      responses:
        "204":
          description: confirmation email sent
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Request deletion of my account
      tags:
      - account
  /api/v1/me/email-change:
    post:
      parameters:
      - description: Proposed new email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.emailChangeRequest'
      responses:
        "204":
          description: verification email sent
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Request a change to my email address
      tags:
      - account
  /api/v1/me/password:
    post:
      parameters:
      - description: New password plus proof of identity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.changePasswordRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TokenPair'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Change my password
      tags:
      - account
  /api/v1/me/reauth/passkey/begin:
    post:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Begin a passkey re-authentication
      tags:
      - account
...
  /api/v1/usage:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.Usage'
        "404":
          description: quotas not configured
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Get my quota usage
      tags:
      - account
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"